+++ NEWS.md
@@ -11,6 +11,7 @@ These functions are ready for use but are still considered experimental and may 
     existing (spread)Sheet.
   * `sheets_append()` adds rows to the data in an existing sheet.
   * `sheets_edit()` writes to a range.
+  * `sheets_flood()` "floods" all cells in a range with the same content.
 
 ### Other new functions and arguments